requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-genai==1.0.0
request==2.31.0
